requirements.txt
supabase
requests
python-dotenv
sentence-transformers
hdbscan
pandas
numpy
openai
google-api-python-client